model/requirements.txt
torch>=2.2.0
torchvision>=0.17.0
timm>=0.9.12
onnx>=1.16.0
onnxruntime>=1.24.1
Pillow>=10.2.0
numpy>=1.26.4
tqdm>=4.66.2
scikit-learn>=1.4.1
matplotlib>=3.8.3
albumentations>=1.4.1
opencv-python>=4.9.0
